requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.3
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==5.2.4
prompt-toolkit==3.0.33
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
vine==5.0.0
wcwidth==0.2.5
